+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Upgrade dependencies
 + Add MSW
 + Add util convert function
++ Add databases options function
 
 ## 1.4.0
 + Upgrade dependencies + refactor Semaphore and Docker